✅ Reverse Integer | Medium | Java | Runtime: 1 ms (99.97%) | Memory: 42.6 MB | 2026-07-23

+++ leetcode/medium/0007-reverse-integer/solution.java
@@ -1,6 +1,6 @@
 class Solution {
     public int reverse(int x) {
-        int reverse = 0;
+        long reverse = 0;
 
         while(x != 0){
             int digit = x % 10;
